+++ src/main/java/website/automate/jenkins/service/ContextParameterResolver.java
@@ -4,11 +4,9 @@ import java.util.HashMap;
 import java.util.Map;
 import java.util.Set;
 
-import hudson.EnvVars;
-
 public class ContextParameterResolver {
 
-    private static final String CONTEXT_PARAMETER_PREFIX = "automate.website.context.";
+    private static final String CONTEXT_PARAMETER_PREFIX = "website.automate.context.";
     
     private static final ContextParameterResolver INSTANCE = new ContextParameterResolver();
     
@@ -16,12 +14,12 @@ public class ContextParameterResolver {
         return INSTANCE;
     }
     
-    public Map<String, String> resolve(EnvVars envVars){
+    public Map<String, String> resolve(Map<String, String> configParameters){
         Map<String, String> contextParameters = new HashMap<String, String>();
-        Set<String> parameterNames = envVars.keySet();
+        Set<String> parameterNames = configParameters.keySet();
         for(String parameterName : parameterNames){
             if(parameterName.startsWith(CONTEXT_PARAMETER_PREFIX)){
-                contextParameters.put(parameterName.substring(CONTEXT_PARAMETER_PREFIX.length()), envVars.get(parameterName));
+                contextParameters.put(parameterName.substring(CONTEXT_PARAMETER_PREFIX.length()), configParameters.get(parameterName));
             }
         }
         return contextParameters;